Temporarily disable subscription ID verification
This is a "chicken and the egg" problem where we need to deploy the
application to be able to create the subscription, but at that time it's
too late to set the environment variable to configure the ID.

There's a way around this I'm sure, but I don't want to spent time on it
right now. It was just implemented as a road block in case a bad actor
tries to exploit the endpoint, but seen as the risk for that is
relatively low it can be postponed for now.