Implement the activity feed
A feed of activities, including the athlete's own and activities from
athletes that are being followed, are presented on the home page.

The feed is implemented as a custom component to avoid having the Index
page growing too large.

An observable stream of updates to activities is implemented in
ActivityService to make the feed reactive and up-to-date.