Remove custom serialization of lists in EF Core
EF Core 8 has support for list types using JSON to serialize the
property under the hood.

SQLite support for this feature is not 100% (e.g. cannot do all kinds of
filtering on JSON properties), but it's better and cleaner than having
custom code to store the list as a comma separated string.

NB! This is a breaking change even though there are no migrations
involved as the serialization format has changed. The challenge feature
has though never been released to prod so it's not an issue.