Add configmap to override appsettings
Mounting the configmap as a file and using it as an extra configuration
source. This way we can override configuration set in the bundled
appsettings.json file.

By mounting the configmap as a file rather than setting env variables we
are able to use .NET's change detection to update the configuration when
the configmap is changed. One use-case for this is the log level which
we can now update on-the-fly and Serilog will pick up on the change.