Enable re-initialization of infinite scrolling
By having the 'HasMore' flag as a parameter to InifiniteScrollingMarker,
we can update the value when the filter changes (as that invalidates the
current data set).

So even if there's no more data to load for the current filter, a change
to the filter (e.g. search text) may produce a different data set and
the infinite scrolling marker should become active again.