Fix invited athletes in edited challenges
The list of invited athletes was not updated when editing a challenge.
This is now added, but I had to remove the challenge creator from the
search result when inviting athletes as it would break some tracking
rules in EF Core. This is not a problem though and perhaps it's better
this way. The creator will always see the challenge (and is able to join
it) anyways, so inviting yourself never made much sense.