Update Strava models and fix activity import
Strava has two different activity models (SummaryActivity and
DetailedActivity) while we were only using one (the detailed one). This
lead to a bug in `StravaClient.GetAthleteActivitiesAsync` where we
deserialized the result to a more detailed type than what was actually
returned which caused activities imported with
`ImportActivitiesForAthlete` to have some missing properties
(e.g. Description).

Now the import gets a list of correctly typed SummaryActivities which it
only uses the ID of to determine if the activity should be updated or
created before fetching the full DetailedActivity. This causes a few
more calls to the Strava API, but we're still way under the rate limit.
The import is just a helper tool anyway and shouldn't be used much.