Replace cronjob with background service
Turns out having a cronjob to do the batch import wasn't a great idea
after all (at least not how it was implemented) as SQLite doesn't like
multiple connections to the same database.

It probably could've worked if the cronjob called an API endpoint
instead, but it was simpler to just replace it with a background
service. Considered bringing in a scheduling library like Quarts,
Hangfire or Coravel, but it seemed overkill for this simple use-case.

Related commit: 105db524c89c