Set culture based on host / domain name
A middleware checks to see if the request host matches any of the
configured hosts in the `CultureByHost` config, and if so sets the
configured culture.

To make different domains work with the authentication flow, the host is
passed as a query parameter in the redirect_uri we pass to Strava. Since
Strava only allows for _one_ redirect domain, we do a subsequent
redirect ourselves if the `host` parameter doesn't match the current
request. This way the auth cookie gets stored on the domain that was
used during login and the user isn't redirected to the "main" domain.