Extend cookie lifetime and refresh access tokens
Extending the cookie lifetime from 1 to 7 days to avoid the dreaded
login page appearing too often.

To make sure we have up-to-date access tokens now that users don't have
to login so often, a middleware `RefreshTokenMiddleware` was added.
This middleware makes sure to refresh the stored access token if it's
close to or already expired whenever a user visits the site.